+++ app/(tab)/quest/page.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const page = () => {
+  return (
+    <div>
+
+    </div>
+  )
+}
+
+export default page
