+++ app/layout.tsx
@@ -1,7 +1,5 @@
 import type { Metadata } from "next";
 import "@/styles/globals.css";
-import Header from "@/components/common/Header";
-import BottomTab from "@/components/common/BottomTab/BottomTab";
 
 export const metadata: Metadata = {
   title: "DoStep",
@@ -16,12 +14,8 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="ko">
-      <body className="relative bg-primary-100 pt-[104px]">
-        <Header />
-        <div className="w-auto mx-[16px] bg-white">
-          {children}
-        </div>
-        <BottomTab />
+      <body>
+        {children}
       </body>
     </html>
   );
