--- app/login/layout.tsx
@@ -1,7 +1,0 @@
-export default function LoginLayout({
-  children,
-}: {
-  children: React.ReactNode;
-}) {
-  return <div className="h-full bg-primary-200">{children}</div>;
-}
