+++ components/common/BottomTab/BottomTab.tsx
@@ -3,7 +3,7 @@ import BottomTabItem from './BottomTabItem'
 
 const BottomTab = () => {
   return (
-    <footer className='w-full h-[85px] fixed bottom-0 z-100 bg-white'>
+    <nav className='w-full h-[85px] fixed bottom-0 z-100 bg-white'>
       <div className='h-[53px] border-t'>
         <div className='my-[6px] flex'>
           <BottomTabItem content={"홈"} />
@@ -13,8 +13,8 @@ const BottomTab = () => {
           <BottomTabItem content={"사용자"} />
         </div>
       </div>
-    </footer>
+    </nav>
   )
 }
 
-export default BottomTab
+export default BottomTab
